Angular Events
Up until this point, we have mainly been reading data in Angular and outputting it to the screen. What if, however, we wanted to actually change some data or respond to user events? That's where Angular event binding comes into play.

Browser based events in Angular are typically tied to methods within their corresponding components. Here is an example of a button that triggers an alert.